requirements.txt
fact-sphere
randfacts
animalapi
discord.py
dotenv-python